Move the scroll view with audio component to shared folder
