+++ app/components/shared/scrollViewWithAudios/AudioControlButton.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { TouchableOpacity, StyleSheet } from 'react-native';
 import Icon from 'react-native-vector-icons/FontAwesome';
 
-import color from '../../themes/color';
-import componentUtil from '../../utils/component_util';
-import { getStyleOfDevice } from '../../utils/responsive_util';
+import color from '../../../themes/color';
+import componentUtil from '../../../utils/component_util';
+import { getStyleOfDevice } from '../../../utils/responsive_util';
 
 const AudioControlButton = (props) => {
-  return <TouchableOpacity onPress={() => props.onPress()} style={styles.button}>
-            <Icon name={props.iconName} size={props.size} color={color.blackColor} />
+  return <TouchableOpacity onPress={() => !!props.onPress && props.onPress()} style={styles.button}>
+            <Icon name={props.iconName} size={props.size} color={color.whiteColor} />
          </TouchableOpacity>
 }
