+++ app/components/shared/scrollViewWithAudios/HeaderAudioControlComponent.js
