+++ app/views/home/HomeView.js
@@ -1,10 +1,13 @@
 import React from 'react';
-import { View, Text } from 'react-native';
+import { View, Text, ScrollView } from 'react-native';
+
+import ScrollViewWithAudioComponent from '../../components/shared/ScrollViewWithAudioComponent';
 
 const HomeView = () => {
   return (
     <View style={{flex: 1}}>
       <Text>Home screen</Text>
+      <ScrollViewWithAudioComponent title="Item list"/>
     </View>
   )
 }
